Rename expand_gene_set_decorator to expand_gene_set
